split routing up into smaller chunks
