+++ routes/messages.js
@@ -7,15 +7,9 @@ const { ensureAuthenticated } = require("../config/auth");
 // set static assets
 router.use("/public", express.static("public"));
 
-// patient record page
+// messages page
 router.get("/", ensureAuthenticated, (req, res) => {
-    // get all patients assigned to doctor
-    db.query(`SELECT * FROM users_patients WHERE doctorID = ?`, [req.user.doctorID], (err, results) => {
-        if (err) console.log(err);
-        res.render("records", { user: req.user, patients: results });
-    });
+    res.render("messages", { user: req.user });
 });
-
-router.post("/", ensureAuthenticated, (req, res) => {});
 
 module.exports = router;
