+++ views/help.ejs
@@ -4,7 +4,7 @@
 		<meta charset="UTF-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<!-- Bootstrap -->
-		<link rel="stylesheet" href="public/css/main.min.css" />
+		<link rel="stylesheet" href="help/public/css/main.min.css" />
 		<script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
 		<script
 			src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js"
@@ -19,9 +19,9 @@
 		<!-- Font Awesome -->
 		<script src="https://kit.fontawesome.com/47ace9daa2.js" crossorigin="anonymous"></script>
 		<!-- Scripts -->
-		<script src="public/scripts/mobile-nav.js" defer></script>
+		<script src="help/public/scripts/mobile-nav.js" defer></script>
 		<!-- Favicon -->
-		<link rel="shortcut icon" type="image/x-icon" href="public/favicon.ico" />
+		<link rel="shortcut icon" type="image/x-icon" href="help/public/favicon.ico" />
 		<title>Adventure Quest | Help</title>
 	</head>
 	<body>
@@ -29,7 +29,7 @@
 		<nav class="navbar sticky-top navbar-expand-lg navbar-dark bg-dark">
 			<div class="container">
 				<a class="navbar-brand" href="/records">
-					<img src="public/img/logo.png" height="30" class="d-inline-block align-top" alt="" loading="lazy" />
+					<img src="help/public/img/logo.png" height="30" class="d-inline-block align-top" alt="" loading="lazy" />
 				</a>
 				<button
 					class="navbar-toggler"
@@ -70,7 +70,7 @@
 					<a class="nav-link" href="/records">Patient Records</a>
 				</li>
 				<li class="nav-item">
-					<a class="nav-link" href="/new">New Patient</a>
+					<a class="nav-link" href="/patient">New Patient</a>
 				</li>
 				<li class="nav-item">
 					<a class="nav-link" href="/quest">New Quest</a>
