+++ views/response.ejs
@@ -4,7 +4,7 @@
 		<meta charset="UTF-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<!-- Bootstrap -->
-		<link rel="stylesheet" href="public/css/main.min.css" />
+		<link rel="stylesheet" href="response/public/css/main.min.css" />
 		<script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
 		<script
 			src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js"
@@ -19,11 +19,11 @@
 		<!-- Font Awesome -->
 		<script src="https://kit.fontawesome.com/47ace9daa2.js" crossorigin="anonymous"></script>
 		<!-- Scripts -->
-		<script src="public/scripts/messages.js" defer></script>
+		<script src="response/public/scripts/messages.js" defer></script>
 		<!-- Favicon -->
-		<link rel="shortcut icon" type="image/x-icon" href="public/favicon.ico" />
+		<link rel="shortcut icon" type="image/x-icon" href="response/public/favicon.ico" />
 		<!-- Custom Styles -->
-		<link rel="stylesheet" href="public/css/message.css" />
+		<link rel="stylesheet" href="response/public/css/message.css" />
 		<title>Adventure Quest | Records</title>
 	</head>
 	<body>
@@ -31,7 +31,7 @@
 		<nav class="navbar sticky-top navbar-expand-lg navbar-dark bg-dark">
 			<div class="container">
 				<a class="navbar-brand" href="#">
-					<img src="public/img/logo.png" height="30" class="d-inline-block align-top" loading="lazy" />
+					<img src="response/public/img/logo.png" height="30" class="d-inline-block align-top" loading="lazy" />
 				</a>
 				<button
 					class="navbar-toggler"
